sql_ddl/pyproject.toml
[tool.poetry]
name = "sql-ddl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
airplanesdk = "^0.3.14"
migra = "^3.0.1663481299"
psycopg2-binary = "^2.9.5"
databaseci = "^6.0.2"
docker = "^6.0.1"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
